+++ backend/back/src/auth/auth.module.ts
@@ -1,10 +1,20 @@
 import { Module } from '@nestjs/common';
+import { JwtModule, JwtService } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
+import config from 'config';
 import { AuthService } from './auth.service';
-import { UserModule } from 'src/user/user.module';
 
 @Module({
-  imports: [UserModule],
-  providers: [AuthService],
+  imports: [
+    JwtModule.register({
+			secret: config.get<string>('jwt.secret'),
+			signOptions: {
+				expiresIn: config.get<string>('jwt.expiresIn'),
+			}
+		}),
+		PassportModule.register({ defaultStrategy: 'jwt'}),
+	],
+  providers: [AuthService, JwtService],
   exports: [AuthService],
 })
 export class AuthModule {}
