+++ backend/back/src/auth/auth.service.ts
@@ -1,19 +1,16 @@
-import { BadGatewayException, BadRequestException, Injectable, InternalServerErrorException, Logger, UnauthorizedException } from '@nestjs/common';
+import { BadGatewayException, BadRequestException, Injectable, Logger, UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
+import * as bcrypt from 'bcrypt';
+import config from 'config';
 import { authenticator } from 'otplib';
 import { toDataURL } from 'qrcode';
 import { IntraTokenDto } from 'src/user/dto/IntraToken.dto';
 import { IntraUserDto } from 'src/user/dto/IntraUser.dto';
 import { User } from 'src/user/user.entity';
-import { UserService } from 'src/user/user.service';
-import { TokenPayload } from './token-payload.entity';
-import * as bcrypt from 'bcrypt';
-import config from 'config';
 
 @Injectable()
 export class AuthService {
 	constructor(
-		private userService: UserService,
 		private jwtService: JwtService,
 	) { }
 
@@ -65,17 +62,18 @@ export class AuthService {
 		return intraToken;
 	}
 
-	async intraSignIn(code: string) {
+	async intraSignIn(code: string) : Promise<IntraUserDto> {
 		const userToken = await this.getTokenFromIntra(code);
 		const userData = await this.getUserInfoFromIntra(userToken);
-		const curUser = await this.userService.getUserById(userData.id);
+		return userData;
+		// const curUser = await this.userService.getUserById(userData.id);
 
-		if (!this.userService.isUserExist(curUser)) {
-			const newUser = await this.userService.addNewUser(userData);
-			return newUser;
-		}
+		// if (!this.userService.isUserExist(curUser)) {
+		// 	const newUser = await this.userService.addNewUser(userData);
+		// 	return newUser;
+		// }
 		// TODO: refreshToken 이 이미 있는경우 처리
-		return curUser;
+		// return curUser;
 	}
 
 	// Save 2fa secret, but twoFactorEnabled value does not change.
@@ -85,22 +83,7 @@ export class AuthService {
 		return { secret, qr: await this.genQrCodeURL(otpAuthUrl) };
 	}
 
-	async toggleTwoFactor(uid: number) {
-		const user = await this.userService.getUserById(uid);
-		if (this.userService.isUserExist(user)) {
-			if (user.twoFactorEnabled) {
-				user.twoFactorEnabled = !user.twoFactorEnabled;
-				await this.userService.updateUser(user);
-				return null;
-			} else {
-				const { secret, qr } = await this.genTwoFactorSecret(user);
-				user.twoFactorSecret = secret;
-				await this.userService.updateUser(user);
-				return qr;
-			}
-		}
-		throw new UnauthorizedException('User Not Found!');
-	}
+
 
 	async genQrCodeURL(otpAuthUrl: string): Promise<{ data: string }> {
 		return toDataURL(otpAuthUrl);
@@ -122,24 +105,8 @@ export class AuthService {
 		const refresh_token = await this.genRefreshToken(userWithoutPw, false);
 		return { access_token, refresh_token };
 	}
-	async logout(user: User){
-		return await this.userService.deleteRefreshToken(user.uid);
-	}
 
-	async validateUser(email: string, password: string) {
-		const user = await this.userService.getUserByEmail(email);
 
-		if (this.userService.isUserExist(user)) {
-			const isMatch = await bcrypt.compare(password, user.password);
-			if (isMatch) {
-				Logger.log(`User(${email}) login success.`);
-				const { password, ...userWithoutPw } = user;
-				return userWithoutPw;
-			}
-			throw new UnauthorizedException('Wrong password!');
-		}
-		throw new UnauthorizedException('User not found!');
-	}
 
 	async genAccessToken(user: Omit<User, 'password'>, twoFactor: boolean) {
 		const payload = {
@@ -162,24 +129,23 @@ export class AuthService {
 		return refreshToken;	
 	}
 
-	async verifyJwtToken(refreshToken: string): Promise<TokenPayload> {
-		try {
-			const payload = await this.jwtService.verify(refreshToken);
-			return payload;
-		}
-		catch (error) {
-			const errMsg = `Failed to verify the refresh token:  ${String(error)}`;
-			Logger.error(errMsg);
-			throw new BadRequestException(errMsg);
-		}
-	}
+	// async verifyJwtToken(refreshToken: string): Promise<TokenPayload> {
+	// 	try {
+	// 		const payload = await this.jwtService.verify(refreshToken);
+	// 		return payload;
+	// 	}
+	// 	catch (error) {
+	// 		const errMsg = `Failed to verify the refresh token:  ${String(error)}`;
+	// 		Logger.error(errMsg);
+	// 		throw new BadRequestException(errMsg);
+	// 	}
+	// }
 
-	async refreshAccessToken(refreshToken: string) {
-		const payload = await this.verifyJwtToken(refreshToken);
-		const user = await this.userService.getUserById(payload.uid);
+	async refreshAccessToken(refresh_token: string, user: User) {
+		const refreshToken = refresh_token.split(' ')[1];
 		const refreshTokenPayload = refreshToken.split('.')[1];
-		const isMatch = await bcrypt.compare(refreshTokenPayload, user?.refreshToken);
-		if (this.userService.isUserExist(user) && isMatch) {
+		const isMatch = await bcrypt.compare(refreshTokenPayload, user.refreshToken);
+		if (isMatch) {
 			return await this.genAccessToken(user, user.twoFactorEnabled);
 		}
 		else {
