+++ backend/back/src/user/user.module.ts
@@ -1,8 +1,18 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { AuthModule } from 'src/auth/auth.module';
+import { UserFollow } from './user-follow.entity';
+import { UserController } from './user.controller';
+import { User } from './user.entity';
 import { UserService } from './user.service';
 
 @Module({
-  imports: [],
+  imports: [
+    TypeOrmModule.forFeature([User]),
+		TypeOrmModule.forFeature([UserFollow]),
+    AuthModule
+  ],
+  controllers: [UserController],
   providers: [UserService],
   exports: [UserService],
 })
