+++ backend/back/src/user/user.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from "@nestjs/common";
+import { BadRequestException, Injectable, Logger, UnauthorizedException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import * as bcrypt from 'bcrypt';
 import config from "config";
+import { AuthService } from "src/auth/auth.service";
 import { Repository } from "typeorm";
 import { IntraUserDto } from "./dto/IntraUser.dto";
 import { PasswordDto } from "./dto/Password.dto";
@@ -13,15 +14,25 @@ export class UserService {
 	constructor(
 		@InjectRepository(User)
 		private userRepository: Repository<User>,
-
 		@InjectRepository(UserFollow)
 		private userFollowRepository: Repository<UserFollow>,
+		private authService: AuthService
 	){}
 
 	async addNewUser(intraUserDto: IntraUserDto): Promise<User> {
 		const user: User = await User.fromIntraUserDto(intraUserDto);
 		await this.userRepository.save(user);
 		return user;
+	}
+
+	async getUserByIntraDto(userData: IntraUserDto){
+		const user = await this.getUserById(userData.id);
+		if (this.isUserExist(user))
+			return user;
+		else {
+			const newUser = await this.addNewUser(userData);
+			return newUser
+		}
 	}
 
 	async getUserById(uid: number) {
@@ -40,8 +51,9 @@ export class UserService {
 	}
 
 	async setUserRefreshToken(user: User, refresh_token: string){
+		const refreshTokenPayload = refresh_token.split('.')[1];
 		const userUpdate = user;
-		userUpdate.refreshToken = await bcrypt.hash(refresh_token, config.get<number>('hash.password.saltOrRounds'));
+		userUpdate.refreshToken = await bcrypt.hash(refreshTokenPayload, config.get<number>('hash.password.saltOrRounds'));
 		await this.updateUser(userUpdate);
 	}
 
@@ -51,15 +63,14 @@ export class UserService {
 		userUpdate.password = userPw;
 		await this.updateUser(userUpdate);
 	}
-
-	async deleteRefreshToken(uid: number) : Promise<number | null>{
+	async deleteRefreshToken(uid: number) {
 		const user = await this.getUserById(uid);
 		if (this.isUserExist(user)){
 			user.refreshToken = null;
 			await this.userRepository.save(user);
-			return uid;
+			return true;
 		}
-		return null;
+		return false;
 	}
 
 	async updateUser(user: User) {
@@ -72,10 +83,12 @@ export class UserService {
 		await this.userRepository.save(user);
 	}
 
-	isUserExist = (user: User | null): user is User => {
-		return user !== null;
+	async logout(user: User){
+		const isLogout : boolean = await this.deleteRefreshToken(user.uid);
+		if (!isLogout)
+			throw new BadRequestException ('RefreshToken Not Deleted');
 	}
-
+ß
 	async follow(curUser: User, userToFollow: User): Promise<void> {
 		const existingFollowing = await this.userFollowRepository.findOne({ where : { fromUserId: curUser.uid, targetToFollowId: userToFollow.uid } });
 		if (existingFollowing) {
@@ -97,4 +110,39 @@ export class UserService {
 		await this.userFollowRepository.remove(existingFollowing);
 	}
 
+	async toggleTwoFactor(uid: number) {
+		const user = await this.getUserById(uid);
+		if (this.isUserExist(user)) {
+			if (user.twoFactorEnabled) {
+				user.twoFactorEnabled = !user.twoFactorEnabled;
+				await this.updateUser(user);
+				return null;
+			} else {
+				const { secret, qr } = await this.authService.genTwoFactorSecret(user);
+				user.twoFactorSecret = secret;
+				await this.updateUser(user);
+				return qr;
+			}
+		}
+		throw new UnauthorizedException('User Not Found!');
+	}
+
+	async validateUser(email: string, password: string) {
+		const user = await this.getUserByEmail(email);
+
+		if (this.isUserExist(user)) {
+			const isMatch = await bcrypt.compare(password, user.password);
+			if (isMatch) {
+				Logger.log(`User(${email}) login success.`);
+				const { password, ...userWithoutPw } = user;
+				return userWithoutPw;
+			}
+			throw new UnauthorizedException('Wrong password!');
+		}
+		throw new UnauthorizedException('User not found!');
+	}
+
+	isUserExist = (user: User | null): user is User => {
+		return user !== null;
+	}
 }
